Refractoring and namespace consolidation
