+++ skypebot/ChatBot.cs
@@ -3,11 +3,11 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
+using skypebot.Services;
 using SKYPE4COMLib;
 
-namespace repostpolice
+namespace skypebot
 {
     public class ChatBot : IChatBot
     {
