+++ skypebot/IChatBot.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using skypebot.Services;
 using SKYPE4COMLib;
 
-namespace repostpolice
+namespace skypebot
 {
     interface IChatBot
     {
