+++ skypebot/Services/IChatBotService.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SKYPE4COMLib;
-
-namespace repostpolice
+﻿namespace skypebot.Services
 {
     public interface IChatBotService
     {
