+++ skypebot/Services/repostpolice/model/UrlHistoryItem.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Mambot
+namespace skypebot.Services.repostpolice.model
 {
     public class UrlHistoryItem
     {
