+++ skypebot/skypebot.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{6AF76FF7-536F-4CEE-876C-88DD3083C03A}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>repostpolice</RootNamespace>
-    <AssemblyName>repostpolice</AssemblyName>
+    <RootNamespace>skypebot</RootNamespace>
+    <AssemblyName>skypebot</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -80,13 +80,14 @@
     <Compile Include="Services\couchpotato\CouchPotatoService.cs" />
     <Compile Include="IChatBot.cs" />
     <Compile Include="Services\couchpotato\model\Movie.cs" />
+    <Compile Include="Services\couchpotato\model\MoviesModel.cs" />
     <Compile Include="Services\IChatBotService.cs" />
-    <Compile Include="model\insult.cs" />
+    <Compile Include="Services\repostpolice\model\insult.cs" />
     <Compile Include="Program.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="model\UrlHistoryItem.cs" />
+    <Compile Include="Services\repostpolice\model\UrlHistoryItem.cs" />
     <Compile Include="Utility\IAuthorizationManager.cs" />
     <EmbeddedResource Include="Migrations\201509072047208__first.resx">
       <DependentUpon>201509072047208__first.cs</DependentUpon>
@@ -102,6 +103,7 @@
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
+      <DesignTime>True</DesignTime>
     </Compile>
     <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
